requirements.txt
Flask==3.0.3
Flask-CORS==4.0.0
supabase==1.0.3
httpx==0.23.3
gotrue==1.2.0
python-dotenv==1.0.0
tensorflow==2.16.1
Pillow==10.2.0
requests==2.31.0
PyJWT==2.10.1
cloudinary==1.41.0